Critical bugfix
CRC values were not correctly calculated in crc.h due to the CRC table
being generated after the CRC was already calculated, which is
definitely a bad thing.

An additional method, killWatchInstructions, was added to
n64hijack.cpp. This NOPs any attempts from the game to override the GS
Watch interrupt. An optional argument was added to disable this
functionality.